I moved GraphicalDistributions into the model to support fixing the graphical algorithm and to avoid circular dependencies between projects. Eventually, we should move all of paireddata and extensions into the statistics project.
